refactor: 重命名 music 插件
